// snapser-sdk/models/AuthUsernamePasswordLoginResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * my-dev-cluster
 * Your custom SDK
 *
 * The version of the OpenAPI document: my-dev-cluster: v2 SDK
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { AuthUser } from './AuthUser';
import {
    AuthUserFromJSON,
    AuthUserFromJSONTyped,
    AuthUserToJSON,
} from './AuthUser';

/**
 * 
 * @export
 * @interface AuthUsernamePasswordLoginResponse
 */
export interface AuthUsernamePasswordLoginResponse {
    /**
     * 
     * @type {AuthUser}
     * @memberof AuthUsernamePasswordLoginResponse
     */
    user?: AuthUser;
}

/**
 * Check if a given object implements the AuthUsernamePasswordLoginResponse interface.
 */
export function instanceOfAuthUsernamePasswordLoginResponse(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function AuthUsernamePasswordLoginResponseFromJSON(json: any): AuthUsernamePasswordLoginResponse {
    return AuthUsernamePasswordLoginResponseFromJSONTyped(json, false);
}

export function AuthUsernamePasswordLoginResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): AuthUsernamePasswordLoginResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'user': !exists(json, 'user') ? undefined : AuthUserFromJSON(json['user']),
    };
}

export function AuthUsernamePasswordLoginResponseToJSON(value?: AuthUsernamePasswordLoginResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'user': AuthUserToJSON(value.user),
    };
}