// snapser-sdk/models/InventoryGetItemsResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * my-dev-cluster
 * Your custom SDK
 *
 * The version of the OpenAPI document: my-dev-cluster: v2 SDK
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { InventoryCatalogItem } from './InventoryCatalogItem';
import {
    InventoryCatalogItemFromJSON,
    InventoryCatalogItemFromJSONTyped,
    InventoryCatalogItemToJSON,
} from './InventoryCatalogItem';

/**
 * 
 * @export
 * @interface InventoryGetItemsResponse
 */
export interface InventoryGetItemsResponse {
    /**
     * Catalog items
     * @type {Array<InventoryCatalogItem>}
     * @memberof InventoryGetItemsResponse
     */
    items?: Array<InventoryCatalogItem>;
    /**
     * Number of items matched by the query
     * @type {number}
     * @memberof InventoryGetItemsResponse
     */
    matchingCatalogItems?: number;
    /**
     * Total number of items in the catalog
     * @type {number}
     * @memberof InventoryGetItemsResponse
     */
    totalCatalogItems?: number;
}

/**
 * Check if a given object implements the InventoryGetItemsResponse interface.
 */
export function instanceOfInventoryGetItemsResponse(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function InventoryGetItemsResponseFromJSON(json: any): InventoryGetItemsResponse {
    return InventoryGetItemsResponseFromJSONTyped(json, false);
}

export function InventoryGetItemsResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): InventoryGetItemsResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'items': !exists(json, 'items') ? undefined : ((json['items'] as Array<any>).map(InventoryCatalogItemFromJSON)),
        'matchingCatalogItems': !exists(json, 'matching_catalog_items') ? undefined : json['matching_catalog_items'],
        'totalCatalogItems': !exists(json, 'total_catalog_items') ? undefined : json['total_catalog_items'],
    };
}

export function InventoryGetItemsResponseToJSON(value?: InventoryGetItemsResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'items': value.items === undefined ? undefined : ((value.items as Array<any>).map(InventoryCatalogItemToJSON)),
        'matching_catalog_items': value.matchingCatalogItems,
        'total_catalog_items': value.totalCatalogItems,
    };
}